feat(wal): offline admin gc/checkpoint/resize and needle enumeration
T8 local face: wal::admin gains run_gc/run_checkpoint/run_resize (flock
guarded; busy volume maps to an EBUSY-style error directing operators to
the remote face), list_needles read-only enumeration via full replay, and
pinned/volume-size/free bytes in the inspect report. WalVolume exposes
wal_gc/wal_checkpoint/wal_admin_stats (compact now runs a real GC cycle);
VolumeEngine dispatches the wal_* admin methods, v1 engine returns
InvalidRequest/None. powerfs-volume CLI gains list-needles/gc/checkpoint/
resize admin subcommands. Backfills C.2 with the final fault-sim run
(2 passed, 1332.55s).

+++ powerfs-core/src/wal/replay.rs
@@ -680,7 +680,7 @@ mod tests {
             let max_lsn = r.index.last_lsn();
             assert!(max_lsn < 40);
             assert_eq!(
-                r.index.needle_count() as u64,
+                r.index.needle_count(),
                 max_lsn,
                 "only contiguous prefix needles present"
             );
